+++ src/Changelog.md
@@ -49,5 +49,9 @@
 
 1. Added functionality for editing a poll by the user who created the poll
 2. Split the vuex into modules, UserAuth and Polls
-3. Added functionality for deleting a choice in a poll
-4. Added functionality to edit a single choice in a single poll
+
+# **DATE**: 02-05-2020
+
+1. Added functionality for deleting a choice in a poll
+2. Added functionality to edit a single choice in a single poll
+3. Made the modal component reusuable and dynamic